Web Password Gen
This is a continuation of my password generator project, this iteration is now a web app - made w/ flask form. 
This page allows the user to enter the required information to generate a password. 

The next stage of this project will be to safely store the generated data, along with relevant tags.
From there, I will integrate retrieve password functionality into the page. 

Once there, I'm going to try and make a main Log-In page 

Any feedback is welcome!